% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfrechet.R
\name{dfrechet}
\alias{dfrechet}
\title{frechet pdf}
\usage{
dfrechet(x, shape, scale = 1)
}
\arguments{
\item{x}{A vector of positive quantiles.}

\item{shape}{A vector of positive numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
frechet cdf.
}
\description{
frechet pdf
}
\examples{
pfrechet(1:5, shape=3, scale=2)
pfrechet(1:5, shape=3, scale=1)

}
