% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhyper2.R
\name{dhyper2}
\alias{dhyper2}
\title{hypergeometric pdf with size, D, N parameterization}
\usage{
dhyper2(x, size, D, N, log = FALSE)
}
\arguments{
\item{x}{A vector of integers between max(0, size + D - N) and pmin(size, D).}

\item{size}{A vector of integers between 1 and N.}

\item{D}{A vector of integers between 1 and N.}

\item{N}{A vector of positive integers.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
hypergeometric cdf.
}
\description{
hypergeometric pdf with size, D, N parameterization
}
\examples{
dhyper2(2, 5, 10, 20)
dhyper2(4, 5, 10, 20)

}
