% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcauchy.R
\name{dlcauchy}
\alias{dlcauchy}
\title{log-cauchy pdf}
\usage{
dlcauchy(x, locationlog = 0, scalelog = 1, log = FALSE)
}
\arguments{
\item{x}{A vector of positive quantiles.}

\item{locationlog}{A vector of numbers.}

\item{scalelog}{A vector of positive numbers.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
log-cauchy pdf.
}
\description{
log-cauchy pdf
}
\examples{
dlcauchy(1:5, locationlog=3, scalelog=2)
dlcauchy(1:5, locationlog=3, scalelog=1)

}
