% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlev.R
\name{dlev}
\alias{dlev}
\title{lev pdf}
\usage{
dlev(q, location = 0, scale = 1)
}
\arguments{
\item{q}{A vector of quantiles.}

\item{location}{A vector of numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
lev cdf.
}
\description{
lev pdf
}
\examples{
dlev(0:5, location=3, scale=2)
dlev(0:5, location=3, scale=1)

}
