% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dllogis.R
\name{dllogis}
\alias{dllogis}
\title{loglogistic pdf}
\usage{
dllogis(x, locationlog = 0, scalelog = 1, log = FALSE)
}
\arguments{
\item{x}{A vector of positive quantiles.}

\item{locationlog}{A vector of numbers.}

\item{scalelog}{A vector of positive numbers.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
loglogistic pdf.
}
\description{
loglogistic pdf
}
\examples{
dllogis(1:5, locationlog=3, scalelog=2)
dllogis(1:5, locationlog=3, scalelog=1)

}
