% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlunif.R
\name{dlunif}
\alias{dlunif}
\title{log-uniform pdf}
\usage{
dlunif(x, location1, location2 = 1)
}
\arguments{
\item{x}{A vector of positive quantiles.}

\item{location1}{A vector of positive numbers.}

\item{location2}{A vector of positive numbers location2 > location1.}
}
\value{
log-uniform pdf
}
\description{
log-uniform pdf
}
\examples{
dlunif(c(.1,1,10), location1=.01, location2=100)
dlunif(c(.1,1,10), location1=.001, location2=1000)

}
