% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnhyper.R
\name{dnhyper}
\alias{dnhyper}
\title{negative hypergeometric pdf with k, D, N parameterization}
\usage{
dnhyper(x, k, D, N, log = FALSE)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{k}{A vector of integers between 1 and D.}

\item{D}{A vector of integers between 1 and N.}

\item{N}{A vector of positive integers.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
negative hypergeometric pdf.
}
\description{
negative hypergeometric pdf with k, D, N parameterization
}
\examples{
dnhyper(2, 5, 10, 20)
dnhyper(2:7, 5, 10, 40)
}
