% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsev.R
\name{dsev}
\alias{dsev}
\title{sev pdf}
\usage{
dsev(x, location = 0, scale = 1)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{location}{A vector of numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
sev pdf.
}
\description{
sev pdf
}
\examples{
dsev(0:5, location=3, scale=2)
dsev(0:5, location=3, scale=1)

}
