% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCenterTI.R
\name{normCenterTI}
\alias{normCenterTI}
\title{Factors for computing a confidence interval for a normal distribution quantile}
\usage{
normCenterTI(conf.level, beta, sample.size)
}
\arguments{
\item{conf.level}{A vector of confidence levels (e.g., 0.90, 0.95, or 0.99).}

\item{beta}{A vector of probabilities, corresponding to the amount of the distribution to be covered.}

\item{sample.size}{A vector of sample size values (corresponding to the sample size providing the data for the construction of the confidence intervals.}
}
\value{
coefficient for computing a confidence interal for a normal distribution quantile.
}
\description{
Factors for computing a confidence interval for a normal distribution quantile
}
\examples{
normCenterTI(0.95, 0.90, 100)
normCenterTI(0.95, 0.99, 100)

}
