% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normQuantileCI.R
\name{normQuantileCI}
\alias{normQuantileCI}
\title{Factors for computing a confidence interval for a normal distribution quantile}
\usage{
normQuantileCI(gamma, pvalue, sample.size, method = "Odeh")
}
\arguments{
\item{gamma}{A vector of confidence levels (e.g., 0.90, 0.95, or 0.99).}

\item{pvalue}{A vector of probabilities, corresponding to the quantile of interest.}

\item{sample.size}{A vector of sample size values (corresponding to the sample size providing the data for the construction of the confidence intervals).}

\item{method}{The method is either Odeh or nctr (which uses the R funcion for a quantile of the noncentral t distribution). The Odeh methods has been shown to be more accurate and it therefore recommended.}
}
\value{
coefficinet for computing a confidence interal for a normal distribution quantile.
}
\description{
Factors for computing a confidence interval for a normal distribution quantile
}
\examples{
normQuantileCI(0.95, 0.90, 100)
normQuantileCI(0.95, 0.99, 100)

}
