% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normTailCI.R
\name{normTailCI}
\alias{normTailCI}
\title{Confidence interval for a normal distribution tail probability}
\usage{
normTailCI(alpha, xk.factor, sample.size)
}
\arguments{
\item{alpha}{A vector of confidence level values (e.g., 0.90, 0.95, or 0.99) for one-sided upper confidence bounds or 1 - (confidence level)  values (e.g., 0.10, 0.05, 0.01) for one-sided lower confidence bounds on a normal distribution probability of being greater than a specified x value. To obtail a two-sided confidence interval use 1-alpha/2 and alpha/2.}

\item{xk.factor}{A vector of studentized values (xbar-x)/s.}

\item{sample.size}{A vector of sample size values (corresponding to the sample size providing the data for the construction of the confidence intervals.}
}
\value{
Endpoint for the one-sided confidence bounds for the probability of being greater than a specified x
}
\description{
Confidence interval for a normal distribution tail probability
}
\examples{
normTailCI(0.025, 1.6, 5)
normTailCI(0.975, 1.6, 5)
normTailCI(0.05, 1.6, 5)
normTailCI(0.95, -1.6, 5)

}
