% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfrechet.R
\name{pfrechet}
\alias{pfrechet}
\title{frechet cdf}
\usage{
pfrechet(q, shape, scale = 1)
}
\arguments{
\item{q}{A vector of positive quantiles.}

\item{shape}{A vector of positive numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
frechet cdf.
}
\description{
frechet cdf
}
\examples{
pfrechet(1:5, shape=3, scale=2)
pfrechet(1:5, shape=3, scale=1)

}
