% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyper2.R
\name{phyper2}
\alias{phyper2}
\title{hypergeometric cdf with size, D, N parameterization}
\usage{
phyper2(q, size, D, N, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{A vector of integers between max(0, size + D - N) and pmin(size, D).}

\item{size}{A vector of integers between 1 and N.}

\item{D}{A vector of integers between 1 and N.}

\item{N}{A vector of positive integers.}

\item{lower.tail}{logical; if TRUE (default), probabilities are Pr[X <= x], otherwise, Pr[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
hypergeometric cdf.
}
\description{
hypergeometric cdf with size, D, N parameterization
}
\examples{
phyper2(2, 5, 10, 20)
phyper2(4, 5, 10, 20)


}
