% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plcauchy.R
\name{plcauchy}
\alias{plcauchy}
\title{log-cauchy cdf}
\usage{
plcauchy(q, locationlog = 0, scalelog = 1, lower.tail = TRUE,
  log.p = FALSE)
}
\arguments{
\item{q}{A vector of positive quantiles.}

\item{locationlog}{A vector of numbers.}

\item{scalelog}{A vector of positive numbers.}

\item{lower.tail}{logical; if TRUE (default), probabilities are Pr[X <= x], otherwise, Pr[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
log-cauchy cdf.
}
\description{
log-cauchy cdf
}
\examples{
plcauchy(1:5, locationlog=3, scalelog=2)
plcauchy(1:5, locationlog=3, scalelog=1)

}
