% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plev.R
\name{plev}
\alias{plev}
\title{lev cdf}
\usage{
plev(q, location = 0, scale = 1)
}
\arguments{
\item{q}{A vector of quantiles.}

\item{location}{A vector of numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
lev cdf.
}
\description{
lev cdf
}
\examples{
plev(0:5, location=3, scale=2)
plev(0:5, location=3, scale=1)

}
