% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pllogis.R
\name{pllogis}
\alias{pllogis}
\title{loglogistic cdf}
\usage{
pllogis(q, locationlog = 0, scalelog = 1, lower.tail = TRUE,
  log.p = FALSE)
}
\arguments{
\item{q}{A vector of positive quantiles.}

\item{locationlog}{A vector of numbers.}

\item{scalelog}{A vector of positive numbers.}

\item{lower.tail}{logical; if TRUE (default), probabilities are Pr[X <= x], otherwise, Pr[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
loglogistic cdf.
}
\description{
loglogistic cdf
}
\examples{
pllogis(1:5, locationlog=3, scalelog=2)
pllogis(1:5, locationlog=3, scalelog=1)

}
