% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plunif.R
\name{plunif}
\alias{plunif}
\title{log-uniform cdf}
\usage{
plunif(q, location1 = 0, location2 = 1)
}
\arguments{
\item{q}{A vector of positive quantiles.}

\item{location1}{A vector of positive numbers.}

\item{location2}{A vector of positive numbers location2 > location1.}
}
\value{
log-uniform cdf
}
\description{
log-uniform cdf
}
\examples{
plunif(c(.1,1,10), location1=.01, location2=100)
plunif(c(.1,1,10), location1=.001, location2=1000)

}
