% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnhyper.R
\name{pnhyper}
\alias{pnhyper}
\title{negative hypergeometric cdf with k, D, N parameterization}
\usage{
pnhyper(q, k, D, N, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{A vector of quantiles.}

\item{k}{A vector of integers between 1 and D.}

\item{D}{A vector of integers between 1 and N.}

\item{N}{A vector of positive integers.}

\item{lower.tail}{logical; if TRUE (default), probabilities are Pr[X <= x], otherwise, Pr[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
negative hypergeometric cdf.
}
\description{
negative hypergeometric cdf with k, D, N parameterization
}
\examples{
pnhyper(2, 5, 10, 20)
pnhyper(2:7, 5, 10, 40)

}
