% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbetabinom.R
\name{qbetabinom}
\alias{qbetabinom}
\title{beta-binomiam quantile}
\usage{
qbetabinom(p, size, shape1, shape2)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{size}{A vector of positive integers.}

\item{shape1}{A vector of numbers greater than 0.}

\item{shape2}{A vector of numbers greater than 0.}
}
\value{
beta-binomiam quantile.
}
\description{
beta-binomiam quantile
}
\examples{
qbetabinom(0.10, 10, 2, 3)
qbetabinom(seq(0.1,0.9,by=0.1), 20, 2, 3)

}
