% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfrechet.R
\name{qfrechet}
\alias{qfrechet}
\title{frechet quantile}
\usage{
qfrechet(p, shape, scale = 1)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{shape}{A vector of positive numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
frechet cdf.
}
\description{
frechet quantile
}
\examples{
qfrechet(seq(0.1,0.9,by=0.1), shape=3, scale=2)
qfrechet(seq(0.1,0.9,by=0.1), shape=3, scale=1)

}
