% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhyper2.R
\name{qhyper2}
\alias{qhyper2}
\title{hypergeometric quantile with size, D, N parameterization}
\usage{
qhyper2(p, size, D, N, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{size}{A vector of integers between 1 and N.}

\item{D}{A vector of integers between 1 and N.}

\item{N}{A vector of positive integers.}

\item{lower.tail}{logical; if TRUE (default), probabilities are Pr[X <= x], otherwise, Pr[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
hypergeometric cdf.
}
\description{
hypergeometric quantile with size, D, N parameterization
}
\examples{
qhyper2(0.1, 5, 10, 20)
qhyper2(0.9, 5, 10, 20)

}
