% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlev.R
\name{qlev}
\alias{qlev}
\title{lev quantile}
\usage{
qlev(p, location = 0, scale = 1)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{location}{A vector of numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
lev cdf.
}
\description{
lev quantile
}
\examples{
qlev(seq(0.1, 0.9, by=0.1), location=3, scale=2)
qlev(seq(0.1, 0.9, by=0.1), location=3, scale=1)

}
