% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qllogis.R
\name{qllogis}
\alias{qllogis}
\title{llogis quantile}
\usage{
qllogis(p, locationlog = 0, scalelog = 1, lower.tail = TRUE,
  log.p = FALSE)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{locationlog}{A vector of numbers.}

\item{scalelog}{A vector of positive numbers.}

\item{lower.tail}{logical; if TRUE (default), probabilities are Pr[X <= x], otherwise, Pr[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
llogis quantile.
}
\description{
llogis quantile
}
\examples{
qllogis(seq(0.1, 0.9, by=0.1), locationlog=3, scalelog=2)
qllogis(seq(0.1, 0.9, by=0.1), locationlog=3, scalelog=1)

}
