% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlunif.R
\name{qlunif}
\alias{qlunif}
\title{log-uniform quantile}
\usage{
qlunif(p, location1, location2 = 1)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{location1}{A vector of positive numbers.}

\item{location2}{A vector of positive numbers location2 > location1.}
}
\value{
log-uniform quantile
}
\description{
log-uniform quantile
}
\examples{
qlunif(seq(0.1, 0.9, by=0.1), location1=.01, location2=100)
qlunif(seq(0.1, 0.9, by=0.1), location1=.001, location2=1000)

}
