% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnhyper.R
\name{qnhyper}
\alias{qnhyper}
\title{negative hypergeometric quantile with k, D, N parameterization}
\usage{
qnhyper(p, k, D, N)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{k}{A vector of integers between 1 and D.}

\item{D}{A vector of integers between 1 and N.}

\item{N}{A vector of positive integers.}
}
\value{
negative hypergeometric quantile.
}
\description{
negative hypergeometric quantile with k, D, N parameterization
}
\examples{
qnhyper(0.1, 5, 10, 20)
qnhyper(seq(0.1,0.9,by=0.1), 5, 10, 40)

}
