% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsev.R
\name{qsev}
\alias{qsev}
\title{sev quantile}
\usage{
qsev(p, location = 0, scale = 1)
}
\arguments{
\item{p}{A vector of probabilities between 0 and 1.}

\item{location}{A vector of numbers.}

\item{scale}{A vector of positive numbers.}
}
\value{
sev cdf.
}
\description{
sev quantile
}
\examples{
qsev(seq(0.1, 0.9, by=0.1), location=3, scale=2)
qsev(seq(0.1, 0.9, by=0.1), location=3, scale=1)

}
